--- .run_publish/simply-robot [clean publishToMavenLocal publishToSonatype closeAndReleaseSonatypeStagingRepository].run.xml
@@ -1,43 +1,0 @@
-<!--
-  ~  Copyright (c) 2022-2022 ForteScarlet <ForteScarlet@163.com>
-  ~
-  ~  本文件是 simply-robot (或称 simple-robot 3.x 、simbot 3.x ) 的一部分。
-  ~
-  ~  simply-robot 是自由软件：你可以再分发之和/或依照由自由软件基金会发布的 GNU 通用公共许可证修改之，无论是版本 3 许可证，还是（按你的决定）任何以后版都可以。
-  ~
-  ~  发布 simply-robot 是希望它能有用，但是并无保障;甚至连可销售和符合某个特定的目的都不保证。请参看 GNU 通用公共许可证，了解详情。
-  ~
-  ~  你应该随程序获得一份 GNU 通用公共许可证的复本。如果没有，请看:
-  ~  https://www.gnu.org/licenses
-  ~  https://www.gnu.org/licenses/gpl-3.0-standalone.html
-  ~  https://www.gnu.org/licenses/lgpl-3.0-standalone.html
-  ~
-  ~
-  -->
-
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="simply-robot [clean publishToMavenLocal publishToSonatype closeAndReleaseSonatypeStagingRepository]" type="GradleRunConfiguration" factoryName="Gradle">
-    <ExternalSystemSettings>
-      <option name="executionName" />
-      <option name="externalProjectPath" value="$PROJECT_DIR$" />
-      <option name="externalSystemIdString" value="GRADLE" />
-      <option name="scriptParameters" value="" />
-      <option name="taskDescriptions">
-        <list />
-      </option>
-      <option name="taskNames">
-        <list>
-          <option value="clean" />
-          <option value="publishToMavenLocal" />
-          <option value="publishToSonatype" />
-          <option value="closeAndReleaseSonatypeStagingRepository" />
-        </list>
-      </option>
-      <option name="vmOptions" />
-    </ExternalSystemSettings>
-    <ExternalSystemDebugServerProcess>true</ExternalSystemDebugServerProcess>
-    <ExternalSystemReattachDebugProcess>true</ExternalSystemReattachDebugProcess>
-    <DebugAllEnabled>false</DebugAllEnabled>
-    <method v="2" />
-  </configuration>
-</component>
