+++ .simbot/icon.png
