--- Roadmap.md
@@ -1,3 +1,0 @@
-# 路线图
-- [ ] 发布`v3.0.0`
-- [ ] 实现多平台(Kotlin/JS and Kotlin/Native)
