+++ buildSrc_/src/main/kotlin_1/simbot.maven-publish.gradle.kts
