+++ gradle.properties
@@ -1,14 +1,13 @@
 kotlin.code.style=official
 kotlin.js.generate.executable.default=false
-kotlin.incremental=true
-kotlin.incremental.js=true
-kotlin.incremental.js.ir=true
-kotlin.mpp.stability.nowarn=true
-kotlin.native.ignoreDisabledTargets=true
-kotlin.daemon.jvmargs=-Xmx2g -Dfile.encoding=UTF-8
+#kotlin.mpp.stability.nowarn=true
+#kotlin.native.ignoreDisabledTargets=true
+#kotlin.daemon.jvmargs=-Xmx2g -Dfile.encoding=UTF-8
 
-org.gradle.workers.max=4
-org.gradle.jvmargs=-Xmx2g
-org.gradle.daemon=true
-org.gradle.caching=true
-org.gradle.parallel=true
+
+
+#org.gradle.workers.max=4
+#org.gradle.jvmargs=-Xmx2g -Dfile.encoding=UTF-8
+org.gradle.daemon=false
+#org.gradle.caching=true
+#org.gradle.parallel=true
