--- simbot-utils/simbot-util-stage-loop/src/commonTest/kotlin/love/forte/simbot/common/stagelooptest/StageLoopTest.kt
@@ -1,17 +1,0 @@
-package love.forte.simbot.common.stagelooptest
-
-import kotlin.test.Test
-
-
-/**
- *
- * @author ForteScarlet
- */
-class StageLoopTest {
-
-    @Test
-    fun loopTest() {
-        // TODO
-    }
-    
-}
