+++ simbot-utils/simbot-util-stage-loop2/src/jsMain/kotlin/love/forte/simbot/common/stageloop/StageLoopJs.kt
