+++ simbot-utils/simbot-util-stage-loop2/src/jvmMain/kotlin/love/forte/simbot/common/stageloop/StageLoopJvm.kt
