+++ simbot-utils/simbot-util-stage-loop2/src/nativeMain/kotlin/love/forte/simbot/common/stageloop/StageLoopNative.kt
